Wavelet transforms and EZW coding allow non powers of 2.
Specifically, the transforms now go as deep as they can without hitting an odd size.
The 1d transform stops transforming when the size is odd.
The 2d transform stops transforming one dimension or the other when the size is odd, but
it continues in the other dimension.

This amounts to computing the level differently (times divisible by two rather than log2 of a power
of 2).  EZW coding has been modified to be aware of the new level scheme, and encodes matrices
transformed with the new bounds.